requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
requests
google-generativeai
agno
faiss-cpu
langchain
pymupdf4llm
python-multipart
openmeteo-requests
requests-cache
retry-requests
rank-bm25
thefuzz
googlesearch-python
pycountry
duckduckgo-search
tavily-python
google-genai
langgraph[draw]
langchain-google-genai
seaborn
matplotlib
